+++ atom/src/atomdict.cpp
@@ -1,10 +1,12 @@
 /*-----------------------------------------------------------------------------
-| Copyright (c) 2014-2019, Nucleic
+| Copyright (c) 2014-2023, Nucleic
 |
 | Distributed under the terms of the BSD 3-Clause License.
 |
 | The full license is in the file LICENSE, distributed with this software.
 |----------------------------------------------------------------------------*/
+#include <iostream>
+#include <sstream>
 #include <cppy/cppy.h>
 #include "atomdict.h"
 #include "packagenaming.h"
@@ -157,7 +159,7 @@ PyObject* AtomDict_setdefault( AtomDict* self, PyObject* args )
 	{
 		return 0;
 	}
-	// Get the dictionary from the dict itself in case it was corced.
+	// Get the dictionary from the dict itself in case it was coerced.
 	return cppy::incref( PyDict_GetItem( pyobject_cast( self ), key ) );
 }
 
@@ -222,6 +224,107 @@ static PyType_Slot AtomDict_Type_slots[] = {
     { Py_tp_methods, void_cast( AtomDict_methods ) },              /* tp_methods */
     { Py_tp_base, void_cast( &PyDict_Type ) },                     /* tp_base */
     { Py_tp_new, void_cast( AtomDict_new ) },                      /* tp_new */
+    { 0, 0 },
+};
+
+
+// DefaultAtomDict
+
+int DefaultAtomDict_clear( DefaultAtomDict* self )
+{
+	Py_CLEAR( self->factory );
+	return AtomDict_clear( atomdict_cast( self ) );
+}
+
+
+int DefaultAtomDict_traverse( DefaultAtomDict* self, visitproc visit, void* arg )
+{
+	Py_VISIT( self->factory );
+	return AtomDict_traverse( atomdict_cast( self ), visit, arg );
+}
+
+
+void DefaultAtomDict_dealloc( DefaultAtomDict* self )
+{
+	cppy::clear( &self->factory );
+	AtomDict_dealloc( atomdict_cast( self ) );
+}
+
+static PyObject* DefaultAtomDict_repr( DefaultAtomDict* self )
+{
+	std::ostringstream ostr;
+    ostr << "defaultdict(";
+	cppy::ptr repr( PyObject_Repr( pyobject_cast( self->factory ) ) );
+	if( !repr )
+	{
+		return 0;
+	}
+	ostr << PyUnicode_AsUTF8( repr.get() );
+	ostr << ", ";
+	repr = PyDict_Type.tp_repr( pyobject_cast( self ) );
+	if( !repr )
+	{
+		return 0;
+	}
+	ostr << PyUnicode_AsUTF8( repr.get() );
+    ostr << ")";
+    return PyUnicode_FromString( ostr.str().c_str() );
+}
+
+static PyObject* DefaultAtomDict_missing( DefaultAtomDict* self, PyObject* args )
+{
+	PyObject* key;
+	if( !PyArg_UnpackTuple( args, "__missing__", 1, 1, &key ) )
+	{
+		return 0;
+	}
+    CAtom* atom = self->dict.pointer->data();
+	if( !atom )
+	{
+		return cppy::runtime_error(
+			"Atom object to which this default dict is not alive anymore, "
+			"so missing value cannot be built."
+		);
+	}
+#if PY_VERSION_HEX >= 0x03090000
+	cppy::ptr value_ptr( PyObject_CallOneArg( self->factory, pyobject_cast( atom ) ) );
+#else
+	cppy::ptr temp( PyTuple_Pack(1, pyobject_cast( atom ) ) );
+	cppy::ptr value_ptr( PyObject_Call( self->factory, temp.get(), 0 ) );
+#endif
+	if( !value_ptr )
+	{
+		return 0;
+	}
+	if( should_validate( atomdict_cast( self ) ) )
+	{
+		// We cannot simply validate the value as it will be re-validated when
+		// it is set which leads to creating a different object.
+        if( AtomDict_ass_subscript( atomdict_cast( self ), key, value_ptr.get() ) < 0 )
+		{
+			return 0;
+		}
+		// Get the dictionary from the dict itself in case it was coerced.
+		value_ptr = cppy::incref( PyDict_GetItem( pyobject_cast( self ), key ) );
+	}
+	return value_ptr.release();
+}
+
+static PyMethodDef DefaultAtomDict_methods[] = {
+	{ "__missing__",
+		( PyCFunction )DefaultAtomDict_missing,
+		METH_VARARGS,
+		"Called when a key is absent from the dictionary" },
+	{ 0 } // sentinel
+};
+
+static PyType_Slot DefaultAtomDict_Type_slots[] = {
+    { Py_tp_dealloc, void_cast( DefaultAtomDict_dealloc ) },       /* tp_dealloc */
+    { Py_tp_traverse, void_cast( DefaultAtomDict_traverse ) },     /* tp_traverse */
+    { Py_tp_clear, void_cast( DefaultAtomDict_clear ) },           /* tp_clear */
+	{ Py_tp_repr, void_cast( DefaultAtomDict_repr ) },             /* tp_repr */
+    { Py_tp_methods, void_cast( DefaultAtomDict_methods ) },       /* tp_methods */
+    /* tp_base cannot be set at this stage */
     { 0, 0 },
 };
 
@@ -303,6 +406,57 @@ bool AtomDict::Ready()
 {
     // The reference will be handled by the module to which we will add the type
 	TypeObject = pytype_cast( PyType_FromSpec( &TypeObject_Spec ) );
+    if( !TypeObject )
+    {
+        return false;
+    }
+    return true;
+}
+
+// Initialize static variables (otherwise the compiler eliminates them)
+PyTypeObject* DefaultAtomDict::TypeObject = NULL;
+
+
+PyType_Spec DefaultAtomDict::TypeObject_Spec = {
+	PACKAGE_TYPENAME( "defaultatomdict" ),      /* tp_name */
+	sizeof( DefaultAtomDict ),                  /* tp_basicsize */
+	0,                                          /* tp_itemsize */
+	Py_TPFLAGS_DEFAULT
+	| Py_TPFLAGS_BASETYPE
+	| Py_TPFLAGS_HAVE_GC
+	| Py_TPFLAGS_HAVE_VERSION_TAG,              /* tp_flags */
+    DefaultAtomDict_Type_slots                  /* slots */
+};
+
+
+PyObject* DefaultAtomDict::New( CAtom* atom, Member* key_validator, Member* value_validator, PyObject* factory)
+{
+    cppy::ptr self( PyDict_Type.tp_new( DefaultAtomDict::TypeObject, 0, 0 ) );
+	if( !self )
+	{
+		return 0;
+	}
+    cppy::xincref( pyobject_cast( key_validator ) );
+    atomdict_cast( self.get() )->m_key_validator = key_validator;
+    cppy::xincref( pyobject_cast( value_validator ) );
+    atomdict_cast( self.get() )->m_value_validator = value_validator;
+    atomdict_cast( self.get() )->pointer = new CAtomPointer( atom );
+    cppy::incref( pyobject_cast( factory ) );
+	// XXX validate we do get a callable taking 0 arg
+    defaultatomdict_cast( self.get() )->factory = factory;
+    return self.release();
+}
+
+
+bool DefaultAtomDict::Ready()
+{
+	// This will work only if we create this type after the standard AtomDict
+    // The reference will be handled by the module to which we will add the type
+	PyObject* bases = PyTuple_New( 1 );
+	PyTuple_SET_ITEM( bases, 0, pyobject_cast( AtomDict::TypeObject ) );
+	TypeObject = pytype_cast(
+		PyType_FromSpecWithBases( &TypeObject_Spec, bases )
+	);
     if( !TypeObject )
     {
         return false;