+++ atom/src/catommodule.cpp
@@ -39,6 +39,10 @@ bool ready_types()
     {
         return false;
     }
+    if( !DefaultAtomDict::Ready() )
+    {
+        return false;
+    }
     if( !AtomSet::Ready() )
     {
         return false;
@@ -93,6 +97,14 @@ bool add_objects( PyObject* mod )
 		return false;
 	}
     atom_dict.release();
+
+    // defaultatomdict
+    cppy::ptr defaultatom_dict( pyobject_cast( DefaultAtomDict::TypeObject ) );
+	if( PyModule_AddObject( mod, "defaultatomdict", defaultatom_dict.get() ) < 0 )
+	{
+		return false;
+	}
+    defaultatom_dict.release();
 
     // atomset
     cppy::ptr atom_set( pyobject_cast( AtomSet::TypeObject ) );