+++ tests/test_atom_from_annotations.py
@@ -10,10 +10,12 @@
 """
 import logging
 import sys
+from collections import defaultdict
 from typing import (
     Any,
     Callable as TCallable,
     ClassVar,
+    DefaultDict as TDefaultDict,
     Dict as TDict,
     Iterable,
     List as TList,
@@ -32,6 +34,7 @@ from atom.api import (
     Bool,
     Bytes,
     Callable,
+    DefaultDict,
     Dict,
     Float,
     Instance,
@@ -198,12 +201,33 @@ def test_union_in_annotation(annotation, member, validate_mode):
         (TSet[int], Set(Int()), 1),
         (TDict[int, int], Dict(), 0),
         (TDict[int, int], Dict(Int(), Int()), 1),
+        (TDefaultDict[int, int], DefaultDict(Int(), Int()), 1),
         (TTuple[int], Tuple(), 0),
         (TTuple[int], Tuple(Int()), 1),
         (TTuple[int, ...], Tuple(Int()), 1),
         (TTuple[int, float], Tuple(), 1),
         (TTuple[tuple, int], Tuple(), 1),
-    ],
+    ]
+    + (
+        [
+            (list[int], List(), 0),
+            (list[int], List(Int()), 1),
+            (list[list[int]], List(List()), 1),
+            (list[list[int]], List(List(Int())), -1),
+            (set[int], Set(), 0),
+            (set[int], Set(Int()), 1),
+            (dict[int, int], Dict(), 0),
+            (dict[int, int], Dict(Int(), Int()), 1),
+            (defaultdict[int, int], DefaultDict(Int(), Int()), 1),
+            (tuple[int], Tuple(), 0),
+            (tuple[int], Tuple(Int()), 1),
+            (tuple[int, ...], Tuple(Int()), 1),
+            (tuple[int, float], Tuple(), 1),
+            (tuple[tuple, int], Tuple(), 1),
+        ]
+        if sys.version_info >= (3, 9)
+        else []
+    ),
 )
 def test_annotated_containers_no_default(annotation, member, depth):
     class A(Atom, use_annotations=True, type_containers=depth):
@@ -222,7 +246,13 @@ def test_annotated_containers_no_default(annotation, member, depth):
             ):
                 assert type(k) is type(mk)
                 assert type(v) is type(mv)
-
+        elif isinstance(member, DefaultDict):
+            for (k, v, f), (mk, mv, mf) in zip(
+                A.a.validate_mode[1:], member.validate_mode[1:]
+            ):
+                assert type(k) is type(mk)
+                assert type(v) is type(mv)
+                assert f(A()) == mf(A())
         else:
             assert type(A.a.item) is type(member.item)  # noqa: E721
             if isinstance(member.item, List):
@@ -244,9 +274,20 @@ def test_annotated_containers_no_default(annotation, member, depth):
         (TList, List, [1]),
         (TSet, Set, {1}),
         (TDict, Dict, {1: 2}),
+        (TDefaultDict, DefaultDict, defaultdict(int, {1: 2})),
         (Optional[Iterable], Instance, None),
         (Type[int], Subclass, int),
-    ],
+    ]
+    + (
+        [
+            (list, List, [1]),
+            (set, Set, {1}),
+            (dict, Dict, {1: 2}),
+            (defaultdict, DefaultDict, defaultdict(int, {1: 2})),
+        ]
+        if sys.version_info >= (3, 9)
+        else []
+    ),
 )
 def test_annotations_with_default(annotation, member, default):
     class A(Atom, use_annotations=True):